requirements.txt
absl-py==2.1.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
autopep8==2.0.4
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.0.1
click==8.1.7
comm==0.1.2
Cython==3.0.5
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.11.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==24.3.25
fqdn==1.5.1
gast==0.5.4
google-pasta==0.2.0
googlesearch-python==1.2.3
grpcio==1.63.0
h5py==3.11.0
idna==3.6
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kaggle==1.5.12
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.3.2
mysql-connector-python==8.3.0
namex==0.0.8
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.3.20
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.14.0
PyMySQL==1.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.11.0
uri-template==1.2.0
urllib3==2.1.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.1
whois==0.9.8
widgetsnbextension==4.0.5
wrapt==1.16.0
